refactor(#17): Layered 아키텍처 의존성 방향 수정 (presentation → domain → data)
- layered-data에서 layered-domain 의존성 제거
- layered-domain에 layered-data 의존성 추가
- Repository 인터페이스(StandingsRepository, SampleRepository) 삭제
- RepositoryModule(@Binds) 삭제
- StandingsRepositoryImpl: 인터페이스 제거, 원시 DTO 반환
- SampleRepositoryImpl: 인터페이스 제거
- GetStandingsUseCase: StandingsRepositoryImpl 직접 주입, DTO→모델 변환 담당
- GetSampleDataUseCase 신규 생성
- SampleViewModel: SampleRepository → GetSampleDataUseCase 교체
