+++ layered/layered-data/build.gradle.kts
@@ -34,7 +34,6 @@ android {
 }
 
 dependencies {
-    implementation(project(":layered:layered-domain"))
     implementation(project(":core:core-network"))
     implementation(libs.hilt.android)
     ksp(libs.hilt.android.compiler)
