--- layered/layered-domain/src/main/java/com/soma2026/lab/layered/domain/repository/SampleRepository.kt
@@ -1,5 +1,0 @@
-package com.soma2026.lab.layered.domain.repository
-
-interface SampleRepository {
-    suspend fun getData(): Result<String>
-}
