--- layered/layered-domain/src/main/java/com/soma2026/lab/layered/domain/repository/StandingsRepository.kt
@@ -1,7 +1,0 @@
-package com.soma2026.lab.layered.domain.repository
-
-import com.soma2026.lab.layered.domain.model.Standing
-
-interface StandingsRepository {
-    suspend fun getStandings(): Result<List<Standing>>
-}
